Make launching of browser optional during startup
Invoking the `helpr` function causes a call to `help.start` which yanks the
user out of the R environment and drops them into a browser.  This behavior is
undesirable if a user wishes to make helpr their default help system by adding
a call to `helpr` in their R startup procedure as it causes the user to be
dumped into a web browser *every* time they launch R and forces them to
navigate back to the R console.

This change adds an optional argument, `launch_browser`, to the `helpr`
function which will suppress the call to `help.start` if set to `TRUE`. This
allows users to call `helpr(lanch_browser=FALSE)` in a startup script and delay
the launching of a browser will be delayed until the first time the user
accesses a help page using `?` or `help`.